airflow/requirements.txt
apache-airflow-task-sdk>=1.1.5
pandas>=2.3.3
numpy>=2.3.5
python-dotenv>=1.2.1
twelvedata>=1.2.25
psycopg2-binary>=2.9.11
optuna>=4.6.0
mlflow>=3.8.1
sqlalchemy>=2.0.45
xgboost>=3.1.3
scikit-learn>=1.8.0